requirements.txt
qrcode==6.0
pillow==5.4.0
xlrd=1.2.0